[refactor] filterOptions fetching로직 끌어올림
filterOptions를 fetching하는 로직을 Issue 컴포넌트에서 하도록 끌어올림. label과 milestone개수를 밖에서 계산 후 주입하기위함.
